# backend/artifacts/openapi-v6.0.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "ClearPath Nexus Engine",
    "version": "6.0.0"
  },
  "paths": {
...
    "/api/v1/provenance/routes/{route_id}": {
      "get": {
        "tags": [
          "provenance"
        ],
        "summary": "Get Route Evidence",
        "description": "Return the stored decision evidence. This endpoint never refetches providers.",
        "operationId": "get_route_evidence_api_v1_provenance_routes__route_id__get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "route_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Route Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RouteEvidenceResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/provenance/routes/{route_id}/evidence-kit": {
      "get": {
        "tags": [
          "provenance"
        ],
        "summary": "Get Route Evidence Kit",
        "description": "Return an integrity-checked stored kit without contacting providers.\n\nAn owned route whose decision snapshot was not captured still receives a\ntruthful 200/UNAVAILABLE package. A missing or foreign route remains a\nnon-disclosing 404.",
        "operationId": "get_route_evidence_kit_api_v1_provenance_routes__route_id__evidence_kit_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "route_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Route Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RouteEvidenceKit"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/provenance/records/{record_id}": {
      "get": {
        "tags": [
          "provenance"
        ],
        "summary": "Get Record Neighborhood",
        "operationId": "get_record_neighborhood_api_v1_provenance_records__record_id__get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "record_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Record Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RecordNeighborhoodResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/provenance/sources": {
      "get": {
        "tags": [
          "provenance"
        ],
        "summary": "List Sources",
        "operationId": "list_sources_api_v1_provenance_sources_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SourceListResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "EvidenceComponent": {
        "properties": {
          "role": {
            "type": "string",
            "title": "Role"
          },
          "label": {
            "type": "string",
            "title": "Label"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "record_ids": {
            "items": {
              "type": "string",
              "format": "uuid"
            },
            "type": "array",
            "title": "Record Ids"
          },
          "explanation": {
            "type": "string",
            "title": "Explanation"
          }
        },
        "type": "object",
        "required": [
          "role",
          "label",
          "status",
          "record_ids",
          "explanation"
        ],
        "title": "EvidenceComponent"
      },
      "EvidenceKitIntegrity": {
        "properties": {
          "verified": {
            "type": "boolean",
            "title": "Verified"
          },
          "record_count": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Record Count"
          },
          "checksum_failures": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Checksum Failures"
          }
        },
        "type": "object",
        "required": [
          "verified",
          "record_count"
        ],
        "title": "EvidenceKitIntegrity"
      },
...
      "FreshnessCounts": {
        "properties": {
          "fresh": {
            "type": "integer",
            "title": "Fresh",
            "default": 0
          },
          "aging": {
            "type": "integer",
            "title": "Aging",
            "default": 0
          },
          "stale": {
            "type": "integer",
            "title": "Stale",
            "default": 0
          },
          "unknown": {
            "type": "integer",
            "title": "Unknown",
            "default": 0
          },
          "not_applicable": {
            "type": "integer",
            "title": "Not Applicable",
            "default": 0
          }
        },
        "type": "object",
        "title": "FreshnessCounts"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "LineageEdgeResponse": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "parent_record_id": {
            "type": "string",
            "format": "uuid",
            "title": "Parent Record Id"
          },
          "child_record_id": {
            "type": "string",
            "format": "uuid",
            "title": "Child Record Id"
          },
          "relationship": {
            "type": "string",
            "title": "Relationship"
          }
        },
        "type": "object",
        "required": [
          "id",
          "parent_record_id",
          "child_record_id",
          "relationship"
        ],
        "title": "LineageEdgeResponse"
      },
...
      "ProvenanceRecordResponse": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "entity_type": {
            "type": "string",
            "title": "Entity Type"
          },
          "entity_key": {
            "type": "string",
            "title": "Entity Key"
          },
          "decision_input_role": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Decision Input Role"
          },
          "canonical_source_type": {
            "type": "string",
            "title": "Canonical Source Type"
          },
          "raw_source_state": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Raw Source State"
          },
          "source": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/SourceBrief"
              },
              {
                "type": "null"
              }
            ]
          },
          "observed_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Observed At"
          },
          "fetched_at": {
            "type": "string",
            "format": "date-time",
            "title": "Fetched At"
          },
          "valid_until": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Valid Until"
          },
          "freshness": {
            "additionalProperties": true,
            "type": "object",
            "title": "Freshness"
          },
          "cache_hit": {
            "type": "boolean",
            "title": "Cache Hit"
          },
          "used_in_decision": {
            "type": "boolean",
            "title": "Used In Decision"
          },
          "excluded_reason": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Excluded Reason"
          },
          "quality": {
            "additionalProperties": true,
            "type": "object",
            "title": "Quality"
          },
          "transform": {
            "additionalProperties": true,
            "type": "object",
            "title": "Transform"
          },
          "decision_impact": {
            "additionalProperties": true,
            "type": "object",
            "title": "Decision Impact"
          },
          "request_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Request Id"
          },
          "checksum": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Checksum"
          },
          "value_summary": {
            "additionalProperties": true,
            "type": "object",
            "title": "Value Summary"
          },
          "metadata": {
            "additionalProperties": true,
            "type": "object",
            "title": "Metadata"
          }
        },
        "type": "object",
        "required": [
          "id",
          "entity_type",
          "entity_key",
          "decision_input_role",
          "canonical_source_type",
          "raw_source_state",
          "source",
          "observed_at",
          "fetched_at",
          "valid_until",
          "freshness",
          "cache_hit",
          "used_in_decision",
          "excluded_reason",
          "quality",
          "transform",
          "decision_impact",
          "request_id",
          "checksum",
          "value_summary",
          "metadata"
        ],
        "title": "ProvenanceRecordResponse"
      },
      "ProvenanceSummary": {
        "properties": {
          "decision_record_id": {
            "type": "string",
            "format": "uuid",
            "title": "Decision Record Id"
          },
          "traceability": {
            "$ref": "#/components/schemas/TraceabilityCounts"
          },
          "freshness": {
            "$ref": "#/components/schemas/FreshnessCounts"
          },
          "source_modes": {
            "$ref": "#/components/schemas/SourceModeCounts"
          },
          "decision_use": {
            "$ref": "#/components/schemas/DecisionUseCounts"
          },
          "warnings": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Warnings"
          }
        },
        "type": "object",
        "required": [
          "decision_record_id",
          "traceability",
          "freshness",
          "source_modes",
          "decision_use"
        ],
        "title": "ProvenanceSummary"
      },
...
      "RecordNeighborhoodResponse": {
        "properties": {
          "record": {
            "$ref": "#/components/schemas/ProvenanceRecordResponse"
          },
          "parents": {
            "items": {
              "$ref": "#/components/schemas/ProvenanceRecordResponse"
            },
            "type": "array",
            "title": "Parents"
          },
          "children": {
            "items": {
              "$ref": "#/components/schemas/ProvenanceRecordResponse"
            },
            "type": "array",
            "title": "Children"
          },
          "edges": {
            "items": {
              "$ref": "#/components/schemas/LineageEdgeResponse"
            },
            "type": "array",
            "title": "Edges"
          }
        },
        "type": "object",
        "required": [
          "record",
          "parents",
          "children",
          "edges"
        ],
        "title": "RecordNeighborhoodResponse"
      },
...
      "RouteEvidenceKit": {
        "properties": {
          "generated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Generated At"
          },
          "format_version": {
            "type": "string",
            "title": "Format Version",
            "default": "clearpath.evidence-kit.v1"
          },
          "kit_status": {
            "type": "string",
            "enum": [
              "HOT",
              "DEGRADED",
              "UNAVAILABLE"
            ],
            "title": "Kit Status"
          },
          "decision_state": {
            "type": "string",
            "enum": [
              "HARD_BLOCKED",
              "HOLD",
              "UNAVAILABLE",
              "READY"
            ],
            "title": "Decision State"
          },
          "reason_codes": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Reason Codes"
          },
          "provider_refetch_required": {
            "type": "boolean",
            "const": false,
            "title": "Provider Refetch Required",
            "default": false
          },
          "manifest": {
            "additionalProperties": true,
            "type": "object",
            "title": "Manifest"
          },
          "manifest_checksum": {
            "type": "string",
            "pattern": "^[0-9a-f]{64}$",
            "title": "Manifest Checksum"
          },
          "integrity": {
            "$ref": "#/components/schemas/EvidenceKitIntegrity"
          },
          "evidence": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/RouteEvidenceResponse"
              },
              {
                "type": "null"
              }
            ]
          },
          "limitations": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Limitations"
          }
        },
        "type": "object",
        "required": [
          "generated_at",
          "kit_status",
          "decision_state",
          "reason_codes",
          "manifest",
          "manifest_checksum",
          "integrity",
          "evidence",
          "limitations"
        ],
        "title": "RouteEvidenceKit",
        "description": "Stored, incident-safe evidence package; never triggers provider refetches."
      },
      "RouteEvidenceResponse": {
        "properties": {
          "route": {
            "additionalProperties": true,
            "type": "object",
            "title": "Route"
          },
          "snapshot": {
            "additionalProperties": true,
            "type": "object",
            "title": "Snapshot"
          },
          "traceability": {
            "$ref": "#/components/schemas/ProvenanceSummary"
          },
          "components": {
            "items": {
              "$ref": "#/components/schemas/EvidenceComponent"
            },
            "type": "array",
            "title": "Components"
          },
          "records": {
            "items": {
              "$ref": "#/components/schemas/ProvenanceRecordResponse"
            },
            "type": "array",
            "title": "Records"
          },
          "edges": {
            "items": {
              "$ref": "#/components/schemas/LineageEdgeResponse"
            },
            "type": "array",
            "title": "Edges"
          },
          "warnings": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Warnings"
          }
        },
        "type": "object",
        "required": [
          "route",
          "snapshot",
          "traceability",
          "components",
          "records",
          "edges",
          "warnings"
        ],
        "title": "RouteEvidenceResponse"
      },
...
      "SourceListResponse": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/SourceBrief"
            },
            "type": "array",
            "title": "Items"
          },
          "total": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Total"
          }
        },
        "type": "object",
        "required": [
          "items",
          "total"
        ],
        "title": "SourceListResponse"
      },
...
      "TraceabilityCounts": {
        "properties": {
          "traced": {
            "type": "integer",
            "title": "Traced"
          },
          "total": {
            "type": "integer",
            "title": "Total"
          },
          "coverage_pct": {
            "type": "number",
            "title": "Coverage Pct"
          }
        },
        "type": "object",
        "required": [
          "traced",
          "total",
          "coverage_pct"
        ],
        "title": "TraceabilityCounts"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}